requirements.txt
streamlit>=1.40.0
openai>=1.40.0
Pillow>=10.0.0
requests>=2.31.0
